requirements.txt
pyttsx3
speechrecognition
sounddevice
numpy
opencv-python
Pillow
mujoco
fastapi
uvicorn[standard]
websockets
pydantic
pydantic-settings
python-multipart
httpx>=0.25.0

# === PHASE 2 DEPENDENCIES ===
# Hugging Face Integration
transformers>=4.30.0
torch>=2.0.0
huggingface-hub>=0.34.4

# Emotion Recognition
mediapipe>=0.10.0

# Machine Learning
scikit-learn>=1.3.0
matplotlib>=3.7.0

# === OPTIONAL DEPENDENCIES ===
# Dépendances audio optionnelles (nécessitent PortAudio)
# pyaudio

# Export docs (PDF via pypandoc; fallback HTML)
pypandoc-binary>=1.11

# === VISION / DÉTECTION ===
ultralytics>=8.0.0

# === WHISPER (ASR) ===
openai-whisper>=20231117

# === DEVELOPMENT TOOLS ===
ruff
black
isort
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-timeout>=2.0.0

# Si besoin de MySQL avec Python 3, installe mysqlclient ou PyMySQL :
# mysqlclient
# PyMySQL